% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furv-dailyQC-climatic-outliers.R
\name{climatic_outliers}
\alias{climatic_outliers}
\title{Climatic outliers test:}
\usage{
climatic_outliers()
}
\description{
considers as outliers all daily maximum and minimum temperature values falling
outside a range with p25 - 3 interquartile ranges (IQR, lower bound) and p75 + 3 interquartile
ranges (upper bound). For daily precipitation values the same range but using 5 IQR.
}
\details{
There are one graphical and one text outputs. The graphical output produces four boxplots
  one for precipitation (RR), one for daily maximum temperature (Tmax), one for daily minimum temperature
  (Tmin) and another one for daily temperature range (DTR).Values flagged as outlier represented by
  round circles. The values identified by these boxplots are listed with a simple text file.
  The file lists the outliers grouped by the element that produced the inclusion of the value
  as suspicious.

  Input: A text file: each line is for one calendar day.
  The order of the columns should be: Year - Month - RR - Tmax - Tmin.
  Without header and NA values = -99.9.

  Output: writes a list with suspicious values (outliers) <climatic_outliers.txt>;
  each line represents one day and the order of columns is: Year - Month -
  RR - Tmax - Tmin - DTR. Grouped by the element that produced the inclusion of the value
  as suspicious
}
