% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furv-dailyQC-duplicate-date.R
\name{duplicate_dates}
\alias{duplicate_dates}
\title{Duplicate dates test:
includes all dates that appears more than once.}
\usage{
duplicate_dates()
}
\description{
Duplicate dates test:
includes all dates that appears more than once.
}
\details{
Input: A text file: each line is for one calendar day.
  The order of the columns should be: Year - Month - daily precipitation
  (RR) - Daily Maximum temperature (Tmax) - Daily Minimum Temperature (Tmin).
  Without header and NA values = -99.9.

  Output: writes a list with duplicated dates, that appear more than
  once. Each line represents one day and the order of columns is: Year - Month -
  RR - Tmax - Tmin.
}
