% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furv-dailyQC-big-errors.R
\name{out_of_range}
\alias{out_of_range}
\title{Big errors test:
determines the daily maximum, minimum and precipitation that exceed a
given threshold.}
\usage{
out_of_range()
}
\description{
Given daily maximum (Tmax) and minimum (Tmin) temperatures and daily precipitation
(RR) values, reports RR values exceeding 200mm and Tmax and Tmin values exceeding
50 degC.The output is a list with the days in which RR or Tmax or Tmin exceeds
these thresholds.
}
\details{
Input: A text file: each line is for one calendar day.
  The order of the columns should be: Year - Month - RR - Tmax - Tmin.
  Without header and NA values = -99.9.

  Output: writes a list with suspicious values (RR > 200 mm or Tmax > 50 degC
  or Tmax < -50 degC or Tmin > 50 degC or Tmin < -50 degC) in <out_of_range.txt>;
  each line represents one day and the order of columns is: Year - Month -
  RR - Tmax - Tmin.
}
