% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furv-dailyQC-rounding_check.R
\name{rounding_check}
\alias{rounding_check}
\title{Rounding problems evolution test:
plots for daily maximum (Tmax) and minimum temperature (Tmin) and
precipitation (RR) histograms showing rounding.}
\usage{
rounding_check()
}
\description{
It plots the values after the decimal point to analyse rounding
problems evolution. It shows how frequently each of the 10 possible
values (.0 to .9) appears.
}
\details{
Input: A text file: each line is for one calendar day.
  The order of the columns should be: Year - Month - RR - Tmax - Tmin.
  Without header and NA values = -99.9.

  Output: .pdf file with three histograms, one for RR > 0mm
  another for Tmax and the third for Tmin.
}
