% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furv-dailyQC-temporal-coherence.R
\name{temporal_coherence}
\alias{temporal_coherence}
\title{Temporal coherence test:
determines those records where daily maximum or minimum temperature differences
with previous day is too large.}
\usage{
temporal_coherence()
}
\description{
Given daily maximum (Tmax) and minimum (Tmin) temperatures values of two consecutive days,
reports those records where the Tmax or Tmin differences with the previous day is >= 20 degC.
#' The output is a list with the days in which Tmax or Tmin of two consecutive days is
>= 20 degC.
}
\details{
Input: A text file: each line is for one calendar day.
  The order of the columns should be: Year - Month - RR - Tmax - Tmin.
  Without header and NA values = -99.9.

  Output: writes two lists (one for Tmax and one for Tmin) with suspicious values
  (Tmax or Tmin of two consecutive days >= 20 degC) <jumpstmax.txt and jumpstmin.txt>;
  each line represents one day and the order of columns is: Year - Month -
  RR - Tmax - Tmin.
}
