% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furv-dailyQC-tolerance-test.R
\name{tolerance_test}
\alias{tolerance_test}
\title{Tolerance test:}
\usage{
tolerance_test()
}
\description{
report on occurrences of 4 or more equal consecutive values in daily
maximum (Tmax) and minimum temperatures (Tmin).
}
\details{
Input: A text file: each line is for one calendar day.
  The order of the columns should be: Year - Month - RR - Tmax - Tmin.
  Without header and NA values = -99.9.

  Output: writes two list (one for Tmax and another for Tmin) with a line
  for each sequence of 4 or more consecutive values. Each line represents
  one day and the order of columns is: Year - Month - Tmax (or Tmin) -
  length of sequence.
}
